improve demo, data loaders
